fix: report the sample rate the hardware actually used (F-2)
_start_streaming computed actual_raw_fs correctly from the interval the
driver writes back, then threw it away:

    self._actual_samplerate = self.config.samplerate

The comment there conflated "hide the oversampling ratio from downstream"
with "hide the actual rate". Those are different things. Hiding the
oversampling ratio is what dividing by the decimation ratio does, and that
ratio is exact. The rate the ADC actually ran at is not something to hide:
sample_interval_us is truncated to whole microseconds, so the achieved rate
is generally not the requested one.

Consequence: every displayed frequency was scaled by requested/actual.

  F_max     requested   raw us   raw actual   true decimated   was reported   error
    200        512        488       2049          512.25            512      -0.05%
   2000       8192         30      33333         8333.25           8192      -1.70%
   5000      16384         15      66667        16666.75          16384      -1.70%
  10000      32768         10     100000        33333.33          32768      -1.70%
  50000     131072          7     142857       142857.00         131072      -8.25%

At F_max=2000 a true 100 Hz tone displayed at 98.3 Hz and a 60 Hz line read
59.0 Hz — which breaks harmonic-family identification, sideband spacing, and
any BPFO/BPFI comparison against a nameplate, i.e. the core diagnostic
activity. It also biased the integration, which uses the labelled f.

Now reported as actual_raw_fs / effective_osr via the new
_report_samplerate(). The value is carried as a float rather than rounded:
the true rate generally is not an integer, and rounding would reintroduce a
smaller version of the same error. VibeSample.samplerate,
ChannelResult.samplerate and the HDF5 'samplerate' attribute are float
accordingly; existing int-valued files still read back fine.

Covered by tests/test_measurement_validity.py::test_f2_*.

+++ src/rev80/collector.py
@@ -27,7 +27,7 @@ def _write_channel_group(h5_grp, ch: int, sample: 'rev80.VibeSample') -> None:
     cg.create_dataset('data', data=np.asarray(sample.data, dtype=np.float64))
     cg.attrs['timestamp']  = sample.timestamp
     cg.attrs['rel_time']   = float(sample.rel_time)
-    cg.attrs['samplerate'] = int(sample.samplerate)
+    cg.attrs['samplerate'] = float(sample.samplerate)
     cg.attrs['status']     = str(sample.status)
     cg.attrs['overflow']   = bool(sample.overflow)
     cg.attrs['degraded']   = bool(sample.degraded)
@@ -884,7 +884,7 @@ class DataCollector:
 
         ts_str     = decode(fg.attrs["timestamp"])
         rel_time   = float(fg.attrs["rel_time"])
-        samplerate = int(fg.attrs["samplerate"])
+        samplerate = float(fg.attrs["samplerate"])
         status     = decode(fg.attrs["status"])
         try:
             timestamp = datetime.fromisoformat(ts_str)
@@ -915,7 +915,7 @@ class DataCollector:
 
         # Sync enabled_channels from frame data and clamp maxfreq to file samplerate
         all_channels: set[int] = set()
-        file_samplerate: int = 0
+        file_samplerate: float = 0.0
         for frame in self.data["frame_cache"]:
             for k, v in frame.items():
                 if isinstance(k, int):
